# os/bluestore: optimize the processing in ExtentMap::encode_some