--- a/src/include/buffer.h
+++ b/src/include/buffer.h
@@ -501,7 +501,7 @@ namespace buffer CEPH_BUFFER_API {
 	  pbl->append(bufferptr(bp, 0, l));
 	  bp.set_length(bp.length() - l);
 	  bp.set_offset(bp.offset() + l);
-	} else {
+	} else if (pbl) {
 	  // we are using pbl's append_buffer
 	  size_t l = pos - pbl->append_buffer.end_c_str();
 	  if (l) {
@@ -515,12 +515,18 @@ namespace buffer CEPH_BUFFER_API {
       friend class list;
 
     public:
+      contiguous_appender(char* pos)
+        : pbl(nullptr),
+          pos(pos),
+          deep(true) {
+      }
+
       ~contiguous_appender() {
 	if (bp.have_raw()) {
 	  // we allocated a new buffer
 	  bp.set_length(pos - bp.c_str());
 	  pbl->append(std::move(bp));
-	} else {
+	} else if (pbl) {
 	  // we are using pbl's append_buffer
 	  size_t l = pos - pbl->append_buffer.end_c_str();
 	  if (l) {