This pull request is dedicated to lower the CPU pressure in BlueStore pipeline's front-end.

Quick & rough performance testing results (BlueStore: `bluestore_debug_omit_block_device_write = true`, FIO: `nr_files=64`, `size=256M`, `bs=4k`, `numjobs=1`) during 4 KiB rand-writes across 4 MiB are attached below. The IO writes have been disabled and single feeding thread (consuming whole physical core) was used to expose the CPU bottleneck we have there.

* `master` (3580d224d7bbe4ff8e258c09c75320f06b6b8e49):
```
bluestore: (groupid=0, jobs=1): err= 0: pid=15575: Wed Mar 15 15:29:05 2017
  write: IOPS=30.6k, BW=119MiB/s (125MB/s)(2385MiB/20001msec)
    clat (usec): min=56, max=4991, avg=2239.62, stdev=1132.92
     lat (usec): min=84, max=5023, avg=2271.69, stdev=1133.25
    clat percentiles (usec):
     |  1.00th=[  213],  5.00th=[  502], 10.00th=[  700], 20.00th=[ 1080],
     | 30.00th=[ 1464], 40.00th=[ 1848], 50.00th=[ 2224], 60.00th=[ 2608],
     | 70.00th=[ 2992], 80.00th=[ 3408], 90.00th=[ 3824], 95.00th=[ 4048],
     | 99.00th=[ 4256], 99.50th=[ 4320], 99.90th=[ 4512], 99.95th=[ 4576],
     | 99.99th=[ 4704]
    lat (usec) : 100=0.18%, 250=1.09%, 500=3.65%, 750=6.41%, 1000=6.54%
    lat (msec) : 2=26.19%, 4=50.17%, 10=5.77%
  cpu          : usr=98.28%, sys=1.54%, ctx=7897, majf=0, minf=7671
  IO depths    : 1=0.1%, 2=0.1%, 4=0.9%, 8=3.4%, 16=13.2%, 32=27.2%, >=64=55.2%
     submit    : 0=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.1%, >=64=100.0%
     issued rwt: total=0,610682,0, short=0,0,0, dropped=0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=128

Run status group 0 (all jobs):
  WRITE: bw=119MiB/s (125MB/s), 119MiB/s-119MiB/s (125MB/s-125MB/s), io=2385MiB (2501MB), run=20001-20001msec

 Performance counter stats for '/home/radek/fio/fio ceph-bluestore.fio':

      36920.625226      task-clock (msec)         #    1.641 CPUs utilized          
           1106127      context-switches          #    0.030 M/sec                  
             57620      cpu-migrations            #    0.002 M/sec                  
             79651      page-faults               #    0.002 M/sec                  
      123128155925      cycles                    #    3.335 GHz                      (30.99%)
   <not supported>      stalled-cycles-frontend  
   <not supported>      stalled-cycles-backend   
       93261852300      instructions              #    0.76  insns per cycle          (38.67%)
       20106565367      branches                  #  544.589 M/sec                    (38.64%)
         262092655      branch-misses             #    1.30% of all branches          (38.86%)
       31064336698      L1-dcache-loads           #  841.382 M/sec                    (23.36%)
        3274540355      L1-dcache-load-misses     #   10.54% of all L1-dcache hits    (19.71%)
        1649746776      LLC-loads                 #   44.684 M/sec                    (15.90%)
         327165672      LLC-load-misses           #   39.66% of all LL-cache hits     (22.96%)
   <not supported>      L1-icache-loads          
        1868473489      L1-icache-load-misses     #   50.608 M/sec                    (30.71%)
       29885263168      dTLB-loads                #  809.446 M/sec                    (28.54%)
         205740699      dTLB-load-misses          #    0.69% of all dTLB cache hits   (17.39%)
         195524816      iTLB-loads                #    5.296 M/sec                    (15.78%)
          37196569      iTLB-load-misses          #   19.02% of all iTLB cache hits   (23.14%)
   <not supported>      L1-dcache-prefetches     
   <not supported>      L1-dcache-prefetch-misses

      22.500588004 seconds time elapsed

```
* after the changset:
```
bluestore: (groupid=0, jobs=1): err= 0: pid=19306: Wed Mar 15 15:45:31 2017
  write: IOPS=40.2k, BW=157MiB/s (165MB/s)(3141MiB/20001msec)
    clat (usec): min=67, max=34994, avg=1754.04, stdev=873.51
     lat (usec): min=86, max=35015, avg=1778.36, stdev=874.59
    clat percentiles (usec):
     |  1.00th=[  171],  5.00th=[  426], 10.00th=[  628], 20.00th=[  924],
     | 30.00th=[ 1192], 40.00th=[ 1480], 50.00th=[ 1752], 60.00th=[ 2024],
     | 70.00th=[ 2320], 80.00th=[ 2608], 90.00th=[ 2896], 95.00th=[ 3056],
     | 99.00th=[ 3216], 99.50th=[ 3280], 99.90th=[ 3376], 99.95th=[ 3472],
     | 99.99th=[13504]
    lat (usec) : 100=0.12%, 250=2.10%, 500=4.27%, 750=7.45%, 1000=8.93%
    lat (msec) : 2=36.06%, 4=41.05%, 10=0.01%, 20=0.01%, 50=0.01%
  cpu          : usr=96.78%, sys=3.00%, ctx=9637, majf=0, minf=119240
  IO depths    : 1=0.1%, 2=0.1%, 4=0.9%, 8=3.1%, 16=10.1%, 32=28.2%, >=64=57.7%
     submit    : 0=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=0.0%, 8=0.1%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=100.0%
     issued rwt: total=0,803973,0, short=0,0,0, dropped=0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=128

Run status group 0 (all jobs):
  WRITE: bw=157MiB/s (165MB/s), 157MiB/s-157MiB/s (165MB/s-165MB/s), io=3141MiB (3293MB), run=20001-20001msec

 Performance counter stats for '/home/radek/fio/fio ceph-bluestore.fio':

      38747.117548      task-clock (msec)         #    1.733 CPUs utilized          
           1150669      context-switches          #    0.030 M/sec                  
             30423      cpu-migrations            #    0.785 K/sec                  
            191266      page-faults               #    0.005 M/sec                  
      129629213751      cycles                    #    3.346 GHz                      (30.78%)
   <not supported>      stalled-cycles-frontend  
   <not supported>      stalled-cycles-backend   
      109579169902      instructions              #    0.85  insns per cycle          (38.57%)
       23565725542      branches                  #  608.193 M/sec                    (38.48%)
         326559443      branch-misses             #    1.39% of all branches          (38.61%)
       36429308177      L1-dcache-loads           #  940.181 M/sec                    (24.04%)
        3449323919      L1-dcache-load-misses     #    9.47% of all L1-dcache hits    (19.81%)
        1533309666      LLC-loads                 #   39.572 M/sec                    (16.35%)
         283070954      LLC-load-misses           #   36.92% of all LL-cache hits     (23.42%)
   <not supported>      L1-icache-loads          
        1843635114      L1-icache-load-misses     #   47.581 M/sec                    (31.07%)
       35484225188      dTLB-loads                #  915.790 M/sec                    (25.19%)
         120743225      dTLB-load-misses          #    0.34% of all dTLB cache hits   (23.88%)
         257682737      iTLB-loads                #    6.650 M/sec                    (15.51%)
          25705049      iTLB-load-misses          #    9.98% of all iTLB cache hits   (22.99%)
   <not supported>      L1-dcache-prefetches     
   <not supported>      L1-dcache-prefetch-misses

      22.355413651 seconds time elapsed
```