@ifed01 2017-03-16
It probably make sense to have object pool for SharedBlob instances too.