@branch-predictor 2017-03-27
On 03/26/2017 07:04 AM, Radoslaw Zarzynski wrote:
> [..]

I think you went too far here. What you're attempting to explain here is 
that we need to simply go deep into assembly-level optimizations, but that's 
not the way to go. We're not writing 4k VR game or RTOS with super-strict 
timing requirements. We use modern CPUs and compilers for a reason, and that 
reason is *not* having to learn about how CPUs work internally. This is 
especially important when you're aware of the fact that each CPU vendor has 
their own architecture (and that differs between CPUs alone, with details 
like presence/size of caches or hyperthreading adding to the subject 
complexity). Finally, you're focusing on x86, and Ceph is happily working on 
ARMs too, making any x86-specific optimizations irrelevant (or even 
counter-productive) on ARMs.
What really hurts Ceph at this point, is memory management, data structures 
used and encoding formats employed. These are easily ones of greatest 
weakpoints (performance-wise) of Ceph and these are easily fixable without 
resorting to assembly-level optimization.

> [..]
> Finally, I must disagree that conditional branches are /always/ expensive. I
> would rather say /it depends/. Many factors should be taken into account:
> whether a branch was mispredicted or not, surrounded by many non-branch
> instructions or not (the tight loop case), direct or indirect, near (without
> crossing the cache line/page boundaries) or far to just name few of them.
> Total costs might vary from negligible (predicted, local and interleaved
> with many other instructions -- narrowed basically to the instructions'
> cost) to devastating (far branches causing cache and ITLB/STLB misses and
> thus page walks, page faults, ...).

That's a prime example of what I had in mind writing above. Instead of 
trying to guess the actual cost of particular branch pattern on particular 
CPU and code layout, just treat each branch as "expensive" - with that being 
an abstract term - and try to avoid it like plague. That way you solve two 
problems (branching costs in general and branching costs on particular 
machines) with reduced development cost.