@aclamk 2017-03-27
@branch-predictor 
1. Assembler in ceph
I agree. Never ever we should resort to this.

2. Data structures
All revolves around detecting how good or how bad is specific data structure in specific context.
We make predictions based on our models of computing. 
In classical model of computing if is always best to have R-B tree if you have more then 10 elements.
In practice we know that if elements are small, then linear search through 100 elements is faster.
This is exactly what we observed in BlueStore. Retrieving data from scattered structures proved abmysally slow.

3. Encoding formats
My i7-4910MQ is able to decode 225000000 varints per second using single thread. 
I talk here about real-life varints recorded on BlueStore testing session.

4. Branches
Not all branches are equal. 
Unconditional ones are not a problem.
Conditional, evaluated early are little problem.
Conditional, evaluated late, but properly guessed are little problem.
Only problem is when branch is mispredicted.

And we cannot avoid logical branches.

As for avoiding machine branches I know of only two techniques:
1) loop unrolling
2) simultaneous evaluation and conditional move

Ad 1) 
Works perfectly if we are sure that our code is executed thousend of times.
If code is executed 1.5 times, then having it smaller usually speeds it.

Ad 2)
Compilers sometimes perform this. 
If not, only way to make sure is to look at assembly.