@rzarzynski 2017-03-27
> What really hurts Ceph at this point, is memory management, data structures
used and encoding formats employed.

@branch-predictor: it's wonderful to see we have agreement here! The maximal throughput of serialization primitives is far less than important if, because of the employed data structures, we are unable to feed them at reasonable rate.

Regards,
Radek