@yuriw 2018-07-05
![](https://github.trello.services/images/mini-trello-icon.png) [wip-yuri3-testing-2018-07-05-1554-mimic](https://trello.com/c/M9epe7bK/336-wip-yuri3-testing-2018-07-05-1554-mimic)